+++ packages/orq-rc/RELEASES.md
@@ -109,3 +109,13 @@ Based on:
 - [python v3.1.0-rc.29] packages/orq-rc
 ### Releases
 - [PyPI v3.1.0-rc.29] https://pypi.org/project/orq-ai-sdk/3.1.0-rc.29 - packages/orq-rc
+
+## 2025-02-04 13:50:08
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.483.0 (2.502.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.1.0-rc.34] packages/orq-rc
+### Releases
+- [PyPI v3.1.0-rc.34] https://pypi.org/project/orq-ai-sdk/3.1.0-rc.34 - packages/orq-rc